requirements.txt
google-generativeai
flask
flask-cors
openpyxl
pandas
numpy
scikit-learn
python-dotenv